requirements.txt
PyQt6
PyQt6-WebEngine
paramiko
pyte
pywinpty
